requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
alembic==1.15.2
psycopg[binary]==3.2.6
pydantic==2.10.6
pydantic-settings==2.8.1
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
email-validator==2.2.0
orjson==3.10.16
httpx==0.28.1
structlog==25.2.0
user-agents==2.2.0
pytest==8.3.5
pytest-asyncio==0.26.0